+++ fabricio/docker/container.py
@@ -79,12 +79,13 @@ class Container(object):
             options=self.options,
         )
 
-    def execute(self, cmd, ignore_errors=False, quiet=True):
+    def execute(self, cmd, ignore_errors=False, quiet=True, use_cache=False):
         command = 'docker exec --tty --interactive {container} {cmd}'
         return fabricio.run(
             command.format(container=self, cmd=cmd),
             ignore_errors=ignore_errors,
             quiet=quiet,
+            use_cache=use_cache,
         )
 
     def start(self):